I added another page which is the account page. From the settngs page I created last week, I edited the account page.
There are images on the settings page which I created last week and linked an account page from the account image. I added a mock account page which included information about the account, username, profile, email as well as any other basic information profile pages have on their platform. I added another image for the user to replace with their own and added placeholders for text to be filled in for the user. I added this feature to the website because it helps the user be able to customize their own app and help each user have their own custom features and personalized account when using the Time Management app. My thought process when making this feature was using each image made from the settings page its own page linked from the image which takes the user to another page for them to access the accounts page. Each week my goal is to work on one of the images and the linked page to create a new page with new features.