I labeled the path with the days of the week.
As a team, we all believe that our project should be aesthetically pleasing. In order to do so, it is crucial for the main page to be easy to use and undertsand as well. When making a model to base the main page off of, we had decided to make a colorful boxed path, each labeled with a day of the week where they can see their assignments or tasks. The day of the week labled on the box, allows the user to see what tasks they have to do or complete on that day allowing our project to be simple and functional. A problem I had ran into when adjusting the code to add the labels was the spacing of the boxes. Whenever I added the labels, the boxes order would be messed up and some boxes would not fit onto the defualt page size. In order to fix this problem I had to adjust the width and height of the boxes.